updated install script

+++ install.sh
@@ -1,16 +1,27 @@
 #!/bin/bash
 
-rm -f ~/.vimrc
-rm -f ~/.ackrc
-rm -f ~/.zshrc
-rm -f ~/.tmux.conf
-rm -rf ~/.config/nvim
+read -p "Do you want to remove old config files?" -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  rm -f ~/.vimrc
+  rm -f ~/.ackrc
+  rm -f ~/.zshrc
+  rm -f ~/.tmux.conf
+  rm -f ~/.global_gitignore
+  rm -rf ~/.config/nvim
+fi
 
+
+# nvim config copy
 ln -s `pwd`/.vimrc  ~/.vimrc
 ln -s `pwd`/nvim ~/.config/nvim
 
 ln -s `pwd`/.zshrc ~/.zshrc
 ln -s `pwd`/.ackrc ~/.ackrc
 ln -s `pwd`/.tmux.conf ~/.tmux.conf
+ln -s `pwd`/.global_gitignore ~/.global_gitignore
+
+git config --global core.excludesfile ~/.global_gitignore
 
 echo 'Done c:'
